Issue #377 - use the VMUSB drivers in place rather than as copies
Incorporate MADC32 support from vmusb/daqconfig directly rather than
a copy.
